created module and class BaseDriver
